Stop bundle states for hydro and sonar
