+++ .vscode/settings.json
@@ -1,0 +1,10 @@
+{
+    "python.autoComplete.extraPaths": [
+        "/home/alexandre/ros_sonia_ws/devel/lib/python3/dist-packages",
+        "/opt/ros/noetic/lib/python3/dist-packages"
+    ],
+    "python.analysis.extraPaths": [
+        "/home/alexandre/ros_sonia_ws/devel/lib/python3/dist-packages",
+        "/opt/ros/noetic/lib/python3/dist-packages"
+    ]
+}
