+++ sonia_flexbe_behaviors/src/sonia_flexbe_behaviors/sonar_end_2_end_sm.py
@@ -8,13 +8,13 @@
 ###########################################################
 
 from flexbe_core import Behavior, Autonomy, OperatableStateMachine, ConcurrencyContainer, PriorityContainer, Logger
+from sonia_mapping_states.start_bundle import start_bundle
+from sonia_mapping_states.start_stop_sonar import start_stop_sonar
+from sonia_mapping_states.stop_sonar_bundle import stop_sonar_bundle
 from sonia_navigation_states.init_trajectory import init_trajectory
 from sonia_navigation_states.manual_add_pose_to_trajectory import manual_add_pose_to_trajectory
 from sonia_navigation_states.send_to_planner import send_to_planner
 from sonia_navigation_states.wait_target_reached import wait_target_reached
-from sonia_mapping_states.start_bundle import start_bundle
-from sonia_mapping_states.start_stop_sonar import start_stop_sonar
-from sonia_mapping_states.stop_bundle import stop_bundle
 # Additional imports can be added inside the following tags
 # [MANUAL_IMPORT]
 
@@ -29,7 +29,8 @@ class sonar_end_2_endSM(Behavior):
 	'''
 	sonar end 2 end
 	'''
-	
+
+
 	def __init__(self):
 		super(sonar_end_2_endSM, self).__init__()
 		self.name = 'sonar_end_2_end'
@@ -119,22 +120,22 @@ class sonar_end_2_endSM(Behavior):
 										transitions={'continue': 'init1'},
 										autonomy={'continue': Autonomy.Off})
 
-			# x:734 y:430
-			OperatableStateMachine.add('stop bundle',
-										stop_bundle(ObstacleID=1, resetBundle=False),
-										transitions={'found': 'stop sonar', 'not_found': 'failed', 'time_out': 'failed'},
-										autonomy={'found': Autonomy.Off, 'not_found': Autonomy.Off, 'time_out': Autonomy.Off})
-
 			# x:569 y:421
 			OperatableStateMachine.add('stop sonar',
 										start_stop_sonar(startStop=True),
 										transitions={'continue': 'init2'},
 										autonomy={'continue': Autonomy.Off})
 
+			# x:734 y:430
+			OperatableStateMachine.add('stop sonar bundle',
+										stop_sonar_bundle(sonarObstacleID=1, resetSonarBundle=False),
+										transitions={'found': 'stop sonar', 'not_found': 'failed', 'time_out': 'failed'},
+										autonomy={'found': Autonomy.Off, 'not_found': Autonomy.Off, 'time_out': Autonomy.Off})
+
 			# x:731 y:326
 			OperatableStateMachine.add('tg',
 										wait_target_reached(timeout=30),
-										transitions={'target_reached': 'stop bundle', 'target_not_reached': 'failed', 'error': 'failed'},
+										transitions={'target_reached': 'stop sonar bundle', 'target_not_reached': 'failed', 'error': 'failed'},
 										autonomy={'target_reached': Autonomy.Off, 'target_not_reached': Autonomy.Off, 'error': Autonomy.Off})
 
 			# x:876 y:531
